+++ ExilenceNextApp/src/components/toolbar/toolbarGridItems/netWorthSessionGridItem/NetWorthSessionGridItem.styles.ts
@@ -1,0 +1,13 @@
+import { Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+
+const useStyles = makeStyles((theme: Theme) => ({
+  iconButton: {
+    padding: theme.spacing(0.5),
+  },
+  sessionArea: {
+    padding: `0 ${theme.spacing(1)}`,
+  },
+}));
+
+export default useStyles;
