+++ ExilenceNextApp/src/components/toolbar/toolbarGridItems/overlayAreaGridItem/OverlayAreaGridItem.styles.ts
@@ -1,0 +1,13 @@
+import { Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+
+const useStyles = makeStyles((theme: Theme) => ({
+  iconButton: {
+    padding: theme.spacing(0.5),
+  },
+  overlayArea: {
+    padding: `0 ${theme.spacing(1)}`,
+  },
+}));
+
+export default useStyles;
