+++ ExilenceNextApp/src/components/toolbar/toolbarGridItems/snapshotAreaGridItem/SnapshotAreaGridItem.styles.ts
@@ -1,0 +1,17 @@
+import { Theme } from '@mui/material';
+import makeStyles from '@mui/styles/makeStyles';
+
+const useStyles = makeStyles((theme: Theme) => ({
+  iconButton: {
+    padding: theme.spacing(0.5),
+  },
+  snapshotBtn: {
+    marginLeft: theme.spacing(1),
+    marginRight: theme.spacing(1),
+  },
+  snapshotArea: {
+    padding: `0 ${theme.spacing(1)}`,
+  },
+}));
+
+export default useStyles;
